pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alphagenome_eval"
version = "0.1.0"
description = "Evaluate AlphaGenome on personal genome prediction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Xinming Tu", email = "tuxm@cs.washington.edu"}
]
keywords = ["genomics", "alphagenome", "personal-genome", "prediction", "atac-seq", "rna-seq"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy",
    "pandas",
    "pysam",
    "scikit-learn",
    "scipy",
    "matplotlib",
    "seaborn",
    "pyyaml",
    "tqdm",
    "joblib",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/mostafavilabuw/alphagenome-personal-eval"
Repository = "https://github.com/mostafavilabuw/alphagenome-personal-eval"

[project.optional-dependencies]
gpu = [
    "torch",
    "borzoi-pytorch",
]
dev = [
    "pytest",
    "pytest-cov",
]
all = [
    "alphagenome_eval[gpu,dev]",
    "numba",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
